@@ -1,9 +1,12 @@
 package lipgloss
 
-import "image/color"
+import (
+	"image/color"
+	"strings"
+)
 
 // Set a value on the underlying rules map.
-func (s *Style) set(key propKey, value interface{}) {
+func (s *Style) set(key propKey, value any) {
 	// We don't allow negative integers on any of our other values, so just keep
 	// them at zero or above. We could use uints instead, but the
 	// conversions are a little tedious, so we're sticking with ints for
@@ -13,6 +16,10 @@ func (s *Style) set(key propKey, value interface{}) {
 		s.fgColor = colorOrNil(value)
 	case backgroundKey:
 		s.bgColor = colorOrNil(value)
+	case underlineColorKey:
+		s.ulColor = colorOrNil(value)
+	case underlineKey:
+		s.ul = value.(Underline)
 	case widthKey:
 		s.width = max(0, value.(int))
 	case heightKey:
@@ -29,6 +36,8 @@ func (s *Style) set(key propKey, value interface{}) {
 		s.paddingBottom = max(0, value.(int))
 	case paddingLeftKey:
 		s.paddingLeft = max(0, value.(int))
+	case paddingCharKey:
+		s.paddingChar = value.(rune)
 	case marginTopKey:
 		s.marginTop = max(0, value.(int))
 	case marginRightKey:
@@ -39,6 +48,8 @@ func (s *Style) set(key propKey, value interface{}) {
 		s.marginLeft = max(0, value.(int))
 	case marginBackgroundKey:
 		s.marginBgColor = colorOrNil(value)
+	case marginCharKey:
+		s.marginChar = value.(rune)
 	case borderStyleKey:
 		s.borderStyle = value.(Border)
 	case borderTopForegroundKey:
@@ -49,6 +60,10 @@ func (s *Style) set(key propKey, value interface{}) {
 		s.borderBottomFgColor = colorOrNil(value)
 	case borderLeftForegroundKey:
 		s.borderLeftFgColor = colorOrNil(value)
+	case borderForegroundBlendKey:
+		s.borderBlendFgColor = value.([]color.Color)
+	case borderForegroundBlendOffsetKey:
+		s.borderForegroundBlendOffset = value.(int)
 	case borderTopBackgroundKey:
 		s.borderTopBgColor = colorOrNil(value)
 	case borderRightBackgroundKey:
@@ -67,6 +82,10 @@ func (s *Style) set(key propKey, value interface{}) {
 		s.tabWidth = value.(int)
 	case transformKey:
 		s.transform = value.(func(string) string)
+	case linkKey:
+		s.link = value.(string)
+	case linkParamsKey:
+		s.linkParams = value.(string)
 	default:
 		if v, ok := value.(bool); ok { //nolint:nestif
 			if v {
@@ -95,6 +114,10 @@ func (s *Style) setFrom(key propKey, i Style) {
 		s.set(foregroundKey, i.fgColor)
 	case backgroundKey:
 		s.set(backgroundKey, i.bgColor)
+	case underlineColorKey:
+		s.set(underlineColorKey, i.ulColor)
+	case underlineKey:
+		s.set(underlineKey, i.ul)
 	case widthKey:
 		s.set(widthKey, i.width)
 	case heightKey:
@@ -111,6 +134,8 @@ func (s *Style) setFrom(key propKey, i Style) {
 		s.set(paddingBottomKey, i.paddingBottom)
 	case paddingLeftKey:
 		s.set(paddingLeftKey, i.paddingLeft)
+	case paddingCharKey:
+		s.set(paddingCharKey, i.paddingChar)
 	case marginTopKey:
 		s.set(marginTopKey, i.marginTop)
 	case marginRightKey:
@@ -121,6 +146,8 @@ func (s *Style) setFrom(key propKey, i Style) {
 		s.set(marginLeftKey, i.marginLeft)
 	case marginBackgroundKey:
 		s.set(marginBackgroundKey, i.marginBgColor)
+	case marginCharKey:
+		s.set(marginCharKey, i.marginChar)
 	case borderStyleKey:
 		s.set(borderStyleKey, i.borderStyle)
 	case borderTopForegroundKey:
@@ -131,6 +158,10 @@ func (s *Style) setFrom(key propKey, i Style) {
 		s.set(borderBottomForegroundKey, i.borderBottomFgColor)
 	case borderLeftForegroundKey:
 		s.set(borderLeftForegroundKey, i.borderLeftFgColor)
+	case borderForegroundBlendKey:
+		s.set(borderForegroundBlendKey, i.borderBlendFgColor)
+	case borderForegroundBlendOffsetKey:
+		s.set(borderForegroundBlendOffsetKey, i.borderForegroundBlendOffset)
 	case borderTopBackgroundKey:
 		s.set(borderTopBackgroundKey, i.borderTopBgColor)
 	case borderRightBackgroundKey:
@@ -153,7 +184,7 @@ func (s *Style) setFrom(key propKey, i Style) {
 	}
 }
 
-func colorOrNil(c interface{}) color.Color {
+func colorOrNil(c any) color.Color {
 	if c, ok := c.(color.Color); ok {
 		return c
 	}
@@ -177,7 +208,31 @@ func (s Style) Italic(v bool) Style {
 // whitespace like margins and padding. To change this behavior set
 // [Style.UnderlineSpaces].
 func (s Style) Underline(v bool) Style {
-	s.set(underlineKey, v)
+	if v {
+		return s.UnderlineStyle(UnderlineSingle)
+	}
+	return s.UnderlineStyle(UnderlineNone)
+}
+
+// UnderlineStyle sets the underline style. This can be used to set the underline
+// to be a single, double, curly, dotted, or dashed line.
+//
+// Note that not all terminal emulators support underline styles. If a style is
+// not supported, it will typically fall back to a single underline but this is
+// not guaranteed. This depends on the terminal emulator being used.
+func (s Style) UnderlineStyle(u Underline) Style {
+	s.set(underlineKey, u)
+	return s
+}
+
+// UnderlineColor sets the color of the underline. By default, the underline
+// will be the same color as the foreground.
+//
+// Note that not all terminal emulators support colored underlines. If color is
+// not supported, it might produce unexpected results. This depends on the
+// terminal emulator being used.
+func (s Style) UnderlineColor(c color.Color) Style {
+	s.set(underlineColorKey, c)
 	return s
 }
 
@@ -320,6 +375,18 @@ func (s Style) PaddingBottom(i int) Style {
 	return s
 }
 
+// PaddingChar sets the character used for padding. This is useful for
+// rendering blocks with a specific character, such as a space or a dot.
+// Example of using [NBSP] as padding to prevent line breaks:
+//
+//	```go
+//	s := lipgloss.NewStyle().PaddingChar(lipgloss.NBSP)
+//	```
+func (s Style) PaddingChar(r rune) Style {
+	s.set(paddingCharKey, r)
+	return s
+}
+
 // ColorWhitespace determines whether or not the background color should be
 // applied to the padding. This is true by default as it's more than likely the
 // desired and expected behavior, but it can be disabled for certain graphic
@@ -387,6 +454,13 @@ func (s Style) MarginBottom(i int) Style {
 // the background color on that style will set the margin color on this style.
 func (s Style) MarginBackground(c color.Color) Style {
 	s.set(marginBackgroundKey, c)
+	return s
+}
+
+// MarginChar sets the character used for the margin. This is useful for
+// rendering blocks with a specific character, such as a space or a dot.
+func (s Style) MarginChar(r rune) Style {
+	s.set(marginCharKey, r)
 	return s
 }
 
@@ -532,6 +606,54 @@ func (s Style) BorderBottomForeground(c color.Color) Style {
 // border.
 func (s Style) BorderLeftForeground(c color.Color) Style {
 	s.set(borderLeftForegroundKey, c)
+	return s
+}
+
+// BorderForegroundBlend sets the foreground colors for the border blend. At least
+// 2 colors are required to use blending, otherwise this will no-op with 0 colors,
+// and pass to BorderForeground with 1 color. This will override all other border
+// foreground colors when used.
+//
+// When providing colors, in most cases (e.g. when all border sides are enabled),
+// you will want to provide a wrapping-set of colors, so the start and end color
+// are either the same, or very similar. For example:
+//
+//	lipgloss.NewStyle().BorderForegroundBlend(
+//		lipgloss.Color("#00FA68"),
+//		lipgloss.Color("#9900FF"),
+//		lipgloss.Color("#ED5353"),
+//		lipgloss.Color("#9900FF"),
+//		lipgloss.Color("#00FA68"),
+//	)
+func (s Style) BorderForegroundBlend(c ...color.Color) Style {
+	if len(c) == 0 {
+		return s
+	}
+
+	// Insufficient colors to use blending, pass to BorderForeground.
+	if len(c) == 1 {
+		return s.BorderForeground(c...)
+	}
+
+	s.set(borderForegroundBlendKey, c)
+	return s
+}
+
+// BorderForegroundBlendOffset sets the border blend offset cells, starting from
+// the top left corner. Value can be positive or negative, and does not need to
+// equal the dimensions of the border region. Direction (when positive) is as
+// follows ("o" is starting point):
+//
+//	  o -------->
+//	  ┌──────────┐
+//	^ │          │ |
+//	| │          │ |
+//	| │          │ |
+//	| │          │ v
+//	  └──────────┘
+//	   <---------
+func (s Style) BorderForegroundBlendOffset(v int) Style {
+	s.set(borderForegroundBlendOffsetKey, v)
 	return s
 }
 
@@ -685,6 +807,21 @@ func (s Style) StrikethroughSpaces(v bool) Style {
 //	fmt.Println(s.Render("raow!") // "RAOW!"
 func (s Style) Transform(fn func(string) string) Style {
 	s.set(transformKey, fn)
+	return s
+}
+
+// Hyperlink sets a hyperlink on a style. This is useful for rendering text that
+// can be clicked on in a terminal emulator that supports hyperlinks.
+//
+// Example:
+//
+//	s := lipgloss.NewStyle().Hyperlink("https://charm.sh")
+//	s := lipgloss.NewStyle().Hyperlink("https://charm.sh", "id=1")
+func (s Style) Hyperlink(link string, params ...string) Style {
+	s.set(linkKey, link)
+	if len(params) > 0 {
+		s.set(linkParamsKey, strings.Join(params, ":"))
+	}
 	return s
 }
 
