@@ -29,7 +29,7 @@ import (
 	"fmt"
 	"sync"
 
-	"github.com/charmbracelet/lipgloss/v2"
+	"charm.land/lipgloss/v2"
 )
 
 // Node defines a node in a tree.
@@ -176,7 +176,7 @@ func (t *Tree) Child(children ...any) *Tree {
 			}
 			t.children = t.children.(NodeChildren).Append(newItem)
 		case Children:
-			for i := 0; i < item.Length(); i++ {
+			for i := range item.Length() {
 				t.children = t.children.(NodeChildren).Append(item.At(i))
 			}
 		case Node:
@@ -212,7 +212,7 @@ func ensureParent(nodes Children, item *Tree) (*Tree, int) {
 	parent := nodes.At(j)
 	switch parent := parent.(type) {
 	case *Tree:
-		for i := 0; i < item.Children().Length(); i++ {
+		for i := range item.Children().Length() {
 			parent.Child(item.children.At(i))
 		}
 		return parent, j
