@@ -4,8 +4,8 @@ import (
 	"image/color"
 	"os"
 
+	"charm.land/lipgloss/v2"
 	"github.com/charmbracelet/colorprofile"
-	"github.com/charmbracelet/lipgloss/v2"
 )
 
 var (
