@@ -9,7 +9,7 @@ package main
 import (
 	"os"
 
-	"github.com/charmbracelet/lipgloss/v2"
+	"charm.land/lipgloss/v2"
 )
 
 func main() {
