@@ -21,7 +21,19 @@ func (s Style) GetItalic() bool {
 // GetUnderline returns the style's underline value. If no value is set false is
 // returned.
 func (s Style) GetUnderline() bool {
-	return s.getAsBool(underlineKey, false)
+	return s.ul != UnderlineNone
+}
+
+// GetUnderlineStyle returns the style's underline style. If no value is set
+// UnderlineNone is returned.
+func (s Style) GetUnderlineStyle() Underline {
+	return s.ul
+}
+
+// GetUnderlineColor returns the style's underline color. If no value is set
+// NoColor{} is returned.
+func (s Style) GetUnderlineColor() color.Color {
+	return s.getAsColor(underlineColorKey)
 }
 
 // GetStrikethrough returns the style's strikethrough value. If no value is set false
@@ -135,6 +147,16 @@ func (s Style) GetPaddingLeft() int {
 	return s.getAsInt(paddingLeftKey)
 }
 
+// GetPaddingChar returns the style's padding character. If no value is set a
+// space (`\u0020`) is returned.
+func (s Style) GetPaddingChar() rune {
+	char := s.getAsRune(paddingCharKey)
+	if char == 0 {
+		return ' '
+	}
+	return char
+}
+
 // GetHorizontalPadding returns the style's left and right padding. Unset
 // values are measured as 0.
 func (s Style) GetHorizontalPadding() int {
@@ -184,6 +206,16 @@ func (s Style) GetMarginBottom() int {
 // returned.
 func (s Style) GetMarginLeft() int {
 	return s.getAsInt(marginLeftKey)
+}
+
+// GetMarginChar returns the style's padding character. If no value is set a
+// space (`\u0020`) is returned.
+func (s Style) GetMarginChar() rune {
+	char := s.getAsRune(marginCharKey)
+	if char == 0 {
+		return ' '
+	}
+	return char
 }
 
 // GetHorizontalMargins returns the style's left and right margins. Unset
@@ -264,6 +296,18 @@ func (s Style) GetBorderLeftForeground() color.Color {
 	return s.getAsColor(borderLeftForegroundKey)
 }
 
+// GetBorderForegroundBlend returns the style's border blend foreground
+// colors. If no value is set, nil is returned.
+func (s Style) GetBorderForegroundBlend() []color.Color {
+	return s.getAsColors(borderForegroundBlendKey)
+}
+
+// GetBorderForegroundBlendOffset returns the style's border blend offset. If no
+// value is set, 0 is returned.
+func (s Style) GetBorderForegroundBlendOffset() int {
+	return s.getAsInt(borderForegroundBlendOffsetKey)
+}
+
 // GetBorderTopBackground returns the style's border top background color. If
 // no value is set NoColor{} is returned.
 func (s Style) GetBorderTopBackground() color.Color {
@@ -301,6 +345,9 @@ func (s Style) GetBorderTopWidth() int {
 // runes of varying widths, the widest rune is returned. If no border exists on
 // the top edge, 0 is returned.
 func (s Style) GetBorderTopSize() int {
+	if s.isBorderStyleSetWithoutSides() {
+		return 1
+	}
 	if !s.getAsBool(borderTopKey, false) {
 		return 0
 	}
@@ -311,6 +358,9 @@ func (s Style) GetBorderTopSize() int {
 // runes of varying widths, the widest rune is returned. If no border exists on
 // the left edge, 0 is returned.
 func (s Style) GetBorderLeftSize() int {
+	if s.isBorderStyleSetWithoutSides() {
+		return 1
+	}
 	if !s.getAsBool(borderLeftKey, false) {
 		return 0
 	}
@@ -321,6 +371,9 @@ func (s Style) GetBorderLeftSize() int {
 // contain runes of varying widths, the widest rune is returned. If no border
 // exists on the left edge, 0 is returned.
 func (s Style) GetBorderBottomSize() int {
+	if s.isBorderStyleSetWithoutSides() {
+		return 1
+	}
 	if !s.getAsBool(borderBottomKey, false) {
 		return 0
 	}
@@ -331,6 +384,9 @@ func (s Style) GetBorderBottomSize() int {
 // contain runes of varying widths, the widest rune is returned. If no border
 // exists on the right edge, 0 is returned.
 func (s Style) GetBorderRightSize() int {
+	if s.isBorderStyleSetWithoutSides() {
+		return 1
+	}
 	if !s.getAsBool(borderRightKey, false) {
 		return 0
 	}
@@ -415,9 +471,34 @@ func (s Style) GetTransform() func(string) string {
 	return s.getAsTransform(transformKey)
 }
 
+// GetHyperlink returns the hyperlink along with its parameters. If no
+// hyperlink is set, empty strings are returned.
+func (s Style) GetHyperlink() (link, params string) {
+	if s.isSet(linkKey) {
+		link = s.link
+	}
+	if s.isSet(linkParamsKey) {
+		params = s.linkParams
+	}
+	return
+}
+
 // Returns whether or not the given property is set.
 func (s Style) isSet(k propKey) bool {
 	return s.props.has(k)
+}
+
+func (s Style) getAsRune(k propKey) rune {
+	if !s.isSet(k) {
+		return 0
+	}
+	switch k { //nolint:exhaustive
+	case paddingCharKey:
+		return s.paddingChar
+	case marginCharKey:
+		return s.marginChar
+	}
+	return 0
 }
 
 func (s Style) getAsBool(k propKey, defaultVal bool) bool {
@@ -425,6 +506,19 @@ func (s Style) getAsBool(k propKey, defaultVal bool) bool {
 		return defaultVal
 	}
 	return s.attrs&int(k) != 0
+}
+
+func (s Style) getAsColors(k propKey) (colors []color.Color) {
+	if !s.isSet(k) {
+		return nil
+	}
+
+	switch k { //nolint:exhaustive
+	case borderForegroundBlendKey:
+		return s.borderBlendFgColor
+	}
+
+	return nil
 }
 
 func (s Style) getAsColor(k propKey) color.Color {
@@ -490,6 +584,8 @@ func (s Style) getAsInt(k propKey) int {
 		return s.marginBottom
 	case marginLeftKey:
 		return s.marginLeft
+	case borderForegroundBlendOffsetKey:
+		return s.borderForegroundBlendOffset
 	case maxWidthKey:
 		return s.maxWidth
 	case maxHeightKey:
@@ -531,6 +627,7 @@ func (s Style) getAsTransform(propKey) func(string) string {
 // line.
 func getLines(s string) (lines []string, widest int) {
 	s = strings.ReplaceAll(s, "\t", "    ")
+	s = strings.ReplaceAll(s, "\r\n", "\n")
 	lines = strings.Split(s, "\n")
 
 	for _, l := range lines {
@@ -541,4 +638,18 @@ func getLines(s string) (lines []string, widest int) {
 	}
 
 	return lines, widest
+}
+
+// isBorderStyleSetWithoutSides returns true if the border style is set but no
+// sides are set. This is used to determine if the border should be rendered by
+// default.
+func (s Style) isBorderStyleSetWithoutSides() bool {
+	var (
+		border    = s.getBorderStyle()
+		topSet    = s.isSet(borderTopKey)
+		rightSet  = s.isSet(borderRightKey)
+		bottomSet = s.isSet(borderBottomKey)
+		leftSet   = s.isSet(borderLeftKey)
+	)
+	return border != noBorder && !(topSet || rightSet || bottomSet || leftSet) //nolint:staticcheck
 }
