@@ -3,7 +3,7 @@ package tree_test
 import (
 	"fmt"
 
-	"github.com/charmbracelet/lipgloss/v2/tree"
+	"charm.land/lipgloss/v2/tree"
 	"github.com/charmbracelet/x/ansi"
 )
 
@@ -27,12 +27,12 @@ func ExampleLeaf_SetHidden() {
 	fmt.Println(tr.String())
 	// Output:
 	//
-	// ├── Foo
-	// ├── Bar
-	// │   ├── Qux
-	// │   └── Quux
-	// │       └── Hello!
-	// └── Baz
+	// ├── Foo
+	// ├── Bar
+	// │   ├── Qux
+	// │   └── Quux
+	// │       └── Hello!
+	// └── Baz
 }
 
 func ExampleNewLeaf() {
@@ -55,14 +55,14 @@ func ExampleNewLeaf() {
 
 	fmt.Println(tr.String())
 	// Output:
-	// ├── Foo
-	// ├── Bar
-	// │   ├── Qux
-	// │   ├── Quux
-	// │   │   └── I am groot
-	// │   │       └── leaves
-	// │   └── Quuux
-	// └── Baz
+	// ├── Foo
+	// ├── Bar
+	// │   ├── Qux
+	// │   ├── Quux
+	// │   │   └── I am groot
+	// │   │       └── leaves
+	// │   └── Quuux
+	// └── Baz
 }
 
 func ExampleLeaf_SetValue() {
@@ -85,13 +85,13 @@ func ExampleLeaf_SetValue() {
 	fmt.Println(ansi.Strip(t.String()))
 	// Output:
 	// ⁜ Makeup
-	// ├── Il Makiage
-	// ├── Fenty Beauty
-	// │   ├── Gloss Bomb Universal Lip Luminizer
-	// │   ╰── Hot Cheeks Velour Blushlighter
-	// ├── Nyx
-	// ├── Mac
-	// ╰── Milk
+	// ├── Il Makiage
+	// ├── Fenty Beauty
+	// │   ├── Gloss Bomb Universal Lip Luminizer
+	// │   ╰── Hot Cheeks Velour Blushlighter
+	// ├── Nyx
+	// ├── Mac
+	// ╰── Milk
 }
 
 // Tree Examples
@@ -113,11 +113,11 @@ func ExampleTree_Hide() {
 
 	fmt.Println(tr.String())
 	// Output:
-	// ├── Foo
-	// ├── Bar
-	// │   ├── Qux
-	// │   └── Quuux
-	// └── Baz
+	// ├── Foo
+	// ├── Bar
+	// │   ├── Qux
+	// │   └── Quuux
+	// └── Baz
 }
 
 func ExampleTree_SetHidden() {
@@ -137,10 +137,10 @@ func ExampleTree_SetHidden() {
 	// Hide a tree after its creation. We'll hide Quux.
 	tr.Children().At(1).Children().At(1).SetHidden(true)
 	// Output:
-	// ├── Foo
-	// ├── Bar
-	// │   ├── Qux
-	// │   └── Quuux
-	// └── Baz
+	// ├── Foo
+	// ├── Bar
+	// │   ├── Qux
+	// │   └── Quuux
+	// └── Baz
 	fmt.Println(tr.String())
 }
